core: merge expresslane_cb into connection::expresslane module
Move ExpresslaneCb trait, ExpresslaneCbData, and ExpresslaneCbType from the
separate expresslane_cb.rs into expresslane.rs. All expresslane-related
public types now live in a single module.
